+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- Add Background Security Improvements introduced in macOS 26.1 to Maintenance → Automatic Updates
+
 ### Changed
 
 ### Fixed